+++ QuickTechSystems/Services/DialogService.cs
@@ -48,9 +48,10 @@ namespace QuickTechSystems.WPF.Services
 
             return await taskCompletionSource.Task;
         }
-        public static async Task<int?> ShowQuantityDialog(Window owner, string productName, int initialQuantity = 1)
+
+        public static async Task<decimal?> ShowQuantityDialog(Window owner, string productName, decimal initialQuantity = 1)
         {
-            return await System.Windows.Application.Current.Dispatcher.InvokeAsync<int?>(() =>
+            return await System.Windows.Application.Current.Dispatcher.InvokeAsync<decimal?>(() =>
             {
                 try
                 {
@@ -79,6 +80,7 @@ namespace QuickTechSystems.WPF.Services
                 }
             });
         }
+
         public async Task<(bool Confirmed, string Value)> ShowTextInputDialogAsync(string title, string prompt, Window owner = null)
         {
             var taskCompletionSource = new TaskCompletionSource<(bool Confirmed, string Value)>();
